refactor(sbmd): extract shared JSValue helpers into SbmdJsUtil
Add SbmdJsUtil (mquickjs namespace) with the stateless JSValue accessors
and exception-string helper that were previously duplicated as anonymous-
namespace helpers across SbmdLoader, SbmdResultExecutor, SbmdBundleLoader,
and MQuickJsRuntime. Each consumer now includes the shared header, and
GetUint32Prop takes an explicit default argument.
